I have added the animations for the snail, gems, end goal and worms and have fully Doxygened and commented the code. I animate the player however I have realise that I only have the sprites for the player in one direction so cant do animate it correctly. As such I have decided to leave the player un-animated. As I need to write up the rest of the paperwork for the assignment and have other assignments looming I have decided to stop work on the code. I feel that the code is complete enough to hand in and whilst I could always add more to it, I need to stop at some point and this seems as good a time as ever. I have built the code in release mode and the exe from it works (if in a folder with the assets and dlls) . I have also regenerated the documentation.

+++ PGGAssignment1SDL/collision.cpp
@@ -101,7 +101,7 @@ void Collision::leftTest(float updatedPosition, int minCurrentAxis, int minOppos
 		}
 	}
 	/*loop through the closest*/
-	for (int i = 0; i < closestIIndex.size(); i++)
+	for (int i = 0; i < (int)closestIIndex.size(); i++)
 	{
 
 		/*check if the type of the tile*/
@@ -174,7 +174,7 @@ void Collision::rightTest(float updatedPosition, int maxCurrentAxis, int minOppo
 		}
 	}
 	/*loop through the closest*/
-	for (int i = 0; i < closestIIndex.size(); i++)
+	for (int i = 0; i < (int)closestIIndex.size(); i++)
 	{
 
 		/*check if the type of the tile*/
@@ -247,7 +247,7 @@ void Collision::downTest(float updatedPosition, int maxCurrentAxis, int minOppos
 		}
 	}
 	/*loop through the closest*/
-	for (int i = 0; i < closestIIndex.size(); i++)
+	for (int i = 0; i < (int)closestIIndex.size(); i++)
 	{
 		/*check if the type of the tile*/
 		switch (map->getType(closestIIndex[i], closestJIndex[i]))
@@ -358,14 +358,14 @@ void Collision::roundingCheck(float updatedPosition, int &minMapPosition, int &m
 	if ((int)(updatedPosition / 32) < updatedPosition / 32)
 	{
 		/*has been rounded down*/
-		minMapPosition = ((updatedPosition - extra) / 32);//rounded down to nearest int
-		maxMapPosition = ((updatedPosition - extra) / 32) + 1;//rounded up to nearest int
+		minMapPosition = (int)((updatedPosition - extra) / 32);//rounded down to nearest int
+		maxMapPosition = (int)((updatedPosition - extra) / 32) + 1;//rounded up to nearest int
 	}
 	else
 	{
 		/*has been rounded up*/
-		minMapPosition = ((updatedPosition - extra) / 32) - 1;//rounded down to nearest int
-		maxMapPosition = ((updatedPosition - extra) / 32);//rounded up to nearest int
+		minMapPosition = (int)((updatedPosition - extra) / 32) - 1;//rounded down to nearest int
+		maxMapPosition = (int)((updatedPosition - extra) / 32);//rounded up to nearest int
 	}	
 }
 
@@ -464,7 +464,7 @@ void Collision::enemyLeftTest(float updatedPosition, int minCurrentAxis, int min
 		}
 	}
 	/*loop through the closest*/
-	for (int i = 0; i < closestIIndex.size(); i++)
+	for (int i = 0; i < (int)closestIIndex.size(); i++)
 	{
 
 		/*check if the type of the tile*/
@@ -524,7 +524,7 @@ void Collision::enemyRightTest(float updatedPosition, int maxCurrentAxis, int mi
 		}
 	}
 	/*loop through the closest*/
-	for (int i = 0; i < closestIIndex.size(); i++)
+	for (int i = 0; i < (int)closestIIndex.size(); i++)
 	{
 
 		/*check if the type of the tile*/
